fix: fixed import issues due to renaming refactor: refactored isPublic field to isPrivate in UserEntity
