+++ src/main/java/az/company/bookappbackend/auth/service/AuthService.java
@@ -76,7 +76,7 @@ public class AuthService {
                 .readingExperience(request.readingExperience())
                 .isPrivate(false)
                 .role(Role.USER)
-                .verified(false)
+                .isVerified(false)
                 .isDeleted(false)
                 .build();
